+++ src/layout/sugiyama.c
@@ -238,15 +238,13 @@ static INLINE igraph_real_t igraph_i_median_4(igraph_real_t x1,
  *
  * </para><para>
  * The Sugiyama layout may introduce "bends" on the edges in order to obtain a
- * visually more pleasing layout. This is achieved by adding dummy nodes to
- * edges spanning more than one layer. The resulting layout assigns coordinates
- * not only to the nodes of the original graph but also to the dummy nodes.
- * The layout algorithm will also return the extended graph with the dummy nodes.
- * An edge in the original graph may either be mapped to a single edge in the
- * extended graph or a \em path that starts and ends in the original
- * source and target vertex and passes through multiple dummy vertices. In
- * such cases, the user may also request the mapping of the edges of the extended
- * graph back to the edges of the original graph.
+ * visually more pleasing layout. The additional control points of the edges are
+ * returned in a separate list of matrices, one matrix per edge in the original
+ * graph. If an edge requires no additional control points, the corresponding
+ * matrix will be empty, otherwise the matrix will contain the coordinates of
+ * the control points, one point per row. When drawing the graph, edges should
+ * be drawn in a way that the curve representing the edge passes through the
+ * control points.
  *
  * </para><para>
  * For more details, see K. Sugiyama, S. Tagawa and M. Toda, "Methods for Visual
@@ -255,21 +253,16 @@ static INLINE igraph_real_t igraph_i_median_4(igraph_real_t x1,
  *
  * \param graph Pointer to an initialized graph object.
  * \param res   Pointer to an initialized matrix object. This will contain
- *              the result and will be resized as needed. The first |V| rows
- *              of the layout will contain the coordinates of the original graph,
- *              the remaining rows contain the positions of the dummy nodes.
- *              Therefore, you can use the result both with \p graph or with
- *              \p extended_graph.
- * \param extd_graph Pointer to an uninitialized graph object or \c NULL.
- *                   The extended graph with the added dummy nodes will be
- *                   returned here. In this graph, each edge points downwards
- *                   to lower layers, spans exactly one layer and the first
- *                   |V| vertices coincide with the vertices of the
- *                   original graph.
- * \param extd_to_orig_eids Pointer to a vector or \c NULL. If not \c NULL, the
- *                          mapping from the edge IDs of the extended graph back
- *                          to the edge IDs of the original graph will be stored
- *                          here.
+ *              the result and will be resized as needed. The coordinates of the
+ *              vertices in the layout will be stored in the rows of the matrix,
+ *              one row per vertex.
+ * \param routing Pointer to an uninitialized list of matrices or \c NULL.
+ *                When not \c NULL, the list will be resized as needed such
+ *                that there will be one matrix for each edge of the graph,
+ *                and the matrix will hold the additional control points that
+ *                the edge must pass through, starting from the source vertex
+ *                of the edge and ending at the target vertex. The matrix will
+ *                have zero rows if an edge does not require control points.
  * \param layers  The layer index for each vertex or \c NULL if the layers should
  *                be determined automatically by igraph.
  * \param hgap  The preferred minimum horizontal gap between vertices in the same
@@ -283,36 +276,38 @@ static INLINE igraph_real_t igraph_i_median_4(igraph_real_t x1,
  *                weights when breaking the cycles.
  * \return Error code.
  */
-igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *res,
-                           igraph_t *extd_graph, igraph_vector_int_t *extd_to_orig_eids,
-                           const igraph_vector_int_t* layers, igraph_real_t hgap, igraph_real_t vgap,
-                           igraph_integer_t maxiter, const igraph_vector_t *weights) {
-    igraph_integer_t i, j, k, l, m, nei;
+igraph_error_t igraph_layout_sugiyama(
+    const igraph_t *graph, igraph_matrix_t *res, igraph_matrix_list_t *routing,
+    const igraph_vector_int_t* layers, igraph_real_t hgap, igraph_real_t vgap,
+    igraph_integer_t maxiter, const igraph_vector_t *weights
+) {
+    igraph_integer_t i, j, k, l, nei;
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
+    igraph_integer_t no_of_edges = igraph_ecount(graph);
     igraph_integer_t comp_idx;
-    igraph_integer_t next_extd_vertex_id = no_of_nodes;
     igraph_bool_t directed = igraph_is_directed(graph);
     igraph_integer_t no_of_components;  /* number of components of the original graph */
     igraph_vector_int_t membership;         /* components of the original graph */
-    igraph_vector_int_t extd_edgelist;   /* edge list of the extended graph */
     igraph_vector_int_t layers_own;  /* layer indices after having eliminated empty layers */
-    igraph_real_t dx = 0, dx2 = 0; /* displacement of the current component on the X axis */
+    igraph_real_t dx = 0;   /* displacement of the current component on the X axis */
     igraph_vector_t layer_to_y; /* mapping from layer indices to final Y coordinates */
+    igraph_matrix_t *control_points;
 
     if (layers && igraph_vector_int_size(layers) != no_of_nodes) {
         IGRAPH_ERROR("layer vector too short or too long", IGRAPH_EINVAL);
     }
 
-    if (extd_graph != 0) {
-        IGRAPH_VECTOR_INT_INIT_FINALLY(&extd_edgelist, 0);
-        if (extd_to_orig_eids != 0) {
-            igraph_vector_int_clear(extd_to_orig_eids);
-        }
-    }
-
     IGRAPH_CHECK(igraph_matrix_resize(res, no_of_nodes, 2));
     IGRAPH_VECTOR_INT_INIT_FINALLY(&membership, 0);
     IGRAPH_VECTOR_INIT_FINALLY(&layer_to_y, 0);
+
+    if (routing != 0) {
+        IGRAPH_CHECK(igraph_matrix_list_resize(routing, no_of_edges));
+        for (i = 0; i < no_of_edges; i++) {
+            control_points = igraph_matrix_list_get_ptr(routing, i);
+            IGRAPH_CHECK(igraph_matrix_resize(control_points, 0, 2));
+        }
+    }
 
     /* 1. Find a feedback arc set if we don't have a layering yet. If we do have
      *    a layering, we can leave all the edges as is as they will be re-oriented
@@ -356,6 +351,7 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
         igraph_integer_t component_size, next_new_vertex_id;
         igraph_vector_int_t old2new_vertex_ids;
         igraph_vector_int_t new2old_vertex_ids;
+        igraph_vector_int_t new_vertex_id_to_edge_id;
         igraph_vector_int_t new_layers;
         igraph_vector_int_t edgelist;
         igraph_vector_int_t neis;
@@ -364,9 +360,11 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
         IGRAPH_VECTOR_INT_INIT_FINALLY(&neis, 0);
         IGRAPH_VECTOR_INT_INIT_FINALLY(&new2old_vertex_ids, no_of_nodes);
         IGRAPH_VECTOR_INT_INIT_FINALLY(&old2new_vertex_ids, no_of_nodes);
+        IGRAPH_VECTOR_INT_INIT_FINALLY(&new_vertex_id_to_edge_id, no_of_nodes);
         IGRAPH_VECTOR_INT_INIT_FINALLY(&new_layers, 0);
 
         igraph_vector_int_fill(&old2new_vertex_ids, -1);
+        igraph_vector_int_fill(&new_vertex_id_to_edge_id, -1);
 
         /* Construct a mapping from the old vertex IDs to the new ones */
         for (i = 0, next_new_vertex_id = 0; i < no_of_nodes; i++) {
@@ -402,16 +400,11 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
                 }
                 if (VECTOR(layers_own)[i] == VECTOR(layers_own)[nei]) {
                     /* Edge goes within the same layer, we don't need this in the
-                     * layered graph, but we need it in the extended graph */
-                    if (extd_graph != 0) {
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, i));
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, nei));
-                        if (extd_to_orig_eids != 0) {
-                            IGRAPH_CHECK(igraph_vector_int_push_back(extd_to_orig_eids, eid));
-                        }
-                    }
+                     * layered graph */
                 } else if (VECTOR(layers_own)[i] > VECTOR(layers_own)[nei]) {
-                    /* Edge goes upwards, we have to flip it */
+                    /* Edge goes upwards, we have to flip it and then later on we need to traverse
+                     * its control points in reverse order. We remember that we need to flip it
+                     * by storing -eid-1 in `new_vertex_id_to_edge_id` instead of eid */
                     IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist,
                                                          VECTOR(old2new_vertex_ids)[nei]));
                     for (l = VECTOR(layers_own)[nei] + 1;
@@ -419,27 +412,10 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
                         IGRAPH_CHECK(igraph_vector_int_push_back(&new_layers, l));
                         IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist, next_new_vertex_id));
                         IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist, next_new_vertex_id++));
+                        IGRAPH_CHECK(igraph_vector_int_push_back(&new_vertex_id_to_edge_id, -eid-1));
                     }
                     IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist,
                                                          VECTOR(old2new_vertex_ids)[i]));
-                    /* Also add the edge to the extended graph if needed, but this time
-                     * with the proper orientation */
-                    if (extd_graph != 0) {
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, i));
-                        next_extd_vertex_id += VECTOR(layers_own)[i] - VECTOR(layers_own)[nei] - 1;
-                        for (l = VECTOR(layers_own)[i] - 1, m = 1;
-                             l > VECTOR(layers_own)[nei]; l--, m++) {
-                            IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, next_extd_vertex_id - m));
-                            IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, next_extd_vertex_id - m));
-                            if (extd_to_orig_eids != 0) {
-                                IGRAPH_CHECK(igraph_vector_int_push_back(extd_to_orig_eids, eid));
-                            }
-                        }
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, nei));
-                        if (extd_to_orig_eids != 0) {
-                            IGRAPH_CHECK(igraph_vector_int_push_back(extd_to_orig_eids, eid));
-                        }
-                    }
                 } else {
                     /* Edge goes downwards */
                     IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist,
@@ -449,25 +425,10 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
                         IGRAPH_CHECK(igraph_vector_int_push_back(&new_layers, l));
                         IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist, next_new_vertex_id));
                         IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist, next_new_vertex_id++));
+                        IGRAPH_CHECK(igraph_vector_int_push_back(&new_vertex_id_to_edge_id, eid));
                     }
                     IGRAPH_CHECK(igraph_vector_int_push_back(&edgelist,
                                                          VECTOR(old2new_vertex_ids)[nei]));
-                    /* Also add the edge to the extended graph */
-                    if (extd_graph != 0) {
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, i));
-                        for (l = VECTOR(layers_own)[i] + 1;
-                             l < VECTOR(layers_own)[nei]; l++) {
-                            IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, next_extd_vertex_id));
-                            IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, next_extd_vertex_id++));
-                            if (extd_to_orig_eids != 0) {
-                                IGRAPH_CHECK(igraph_vector_int_push_back(extd_to_orig_eids, eid));
-                            }
-                        }
-                        IGRAPH_CHECK(igraph_vector_int_push_back(&extd_edgelist, nei));
-                        if (extd_to_orig_eids != 0) {
-                            IGRAPH_CHECK(igraph_vector_int_push_back(extd_to_orig_eids, eid));
-                        }
-                    }
                 }
             }
         }
@@ -478,6 +439,10 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
             igraph_matrix_t layout;
             igraph_i_layering_t layering;
             igraph_t subgraph;
+            igraph_integer_t eid;
+            igraph_real_t max_x;
+            igraph_bool_t flipped;
+            igraph_matrix_t *control_points;
 
             IGRAPH_CHECK(igraph_matrix_init(&layout, next_new_vertex_id, 2));
             IGRAPH_FINALLY(igraph_matrix_destroy, &layout);
@@ -507,29 +472,60 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
             IGRAPH_CHECK(igraph_i_layout_sugiyama_place_nodes_horizontally(&subgraph, &layout,
                          &layering, hgap, component_size));
 
-            /* Re-assign rows into the result matrix, and at the same time, */
+            /* Arrange rows of the layout into the result matrix, and at the same time, */
             /* adjust dx so that the next component does not overlap this one */
-            j = next_new_vertex_id - component_size;
-            k = igraph_matrix_nrow(res);
-            IGRAPH_CHECK(igraph_matrix_add_rows(res, j));
-            dx2 = dx;
+
+            /* First we arrange the "real" vertices */
             for (i = 0; i < component_size; i++) {
-                l = VECTOR(new2old_vertex_ids)[i];
-                MATRIX(*res, l, 0) = MATRIX(layout, i, 0) + dx;
-                MATRIX(*res, l, 1) = VECTOR(layer_to_y)[(igraph_integer_t) MATRIX(layout, i, 1)];
-                if (dx2 < MATRIX(*res, l, 0)) {
-                    dx2 = MATRIX(*res, l, 0);
-                }
-            }
-            for (i = component_size; i < next_new_vertex_id; i++) {
+                k = VECTOR(new2old_vertex_ids)[i];
                 MATRIX(*res, k, 0) = MATRIX(layout, i, 0) + dx;
                 MATRIX(*res, k, 1) = VECTOR(layer_to_y)[(igraph_integer_t) MATRIX(layout, i, 1)];
-                if (dx2 < MATRIX(*res, k, 0)) {
-                    dx2 = MATRIX(*res, k, 0);
-                }
-                k++;
             }
-            dx = dx2 + hgap;
+
+            /* Next we arrange the "dummy" vertices that become control points in the
+             * routing matrix list. Note that the dummy vertices were added in a way that
+             * the vertices that will become control points on an edge with a particular ID
+             * are always consecutive in `new_vertex_id_to_edge_id` so we can count the number
+             * of waypoints in an edg easily */
+            if (routing) {
+                IGRAPH_CHECK(igraph_vector_int_push_back(&new_vertex_id_to_edge_id, -1));   /* sentinel */
+                while (i < next_new_vertex_id) {
+                    eid = VECTOR(new_vertex_id_to_edge_id)[i];
+
+                    /* Find out how many control points we have for this edge */
+                    for (j = i; VECTOR(new_vertex_id_to_edge_id)[j] == eid; j++);
+
+                    /* Is this edge flipped in the layered graph? If so, recover the original eid */
+                    flipped = eid < 0;
+                    if (flipped) {
+                        eid = -eid-1;
+                    }
+
+                    control_points = igraph_matrix_list_get_ptr(routing, eid);
+                    IGRAPH_CHECK(igraph_matrix_resize(control_points, j - i, 2));
+
+                    if (flipped) {
+                        for (k = j - i - 1; i < j; k--, i++) {
+                            MATRIX(*control_points, k, 0) = MATRIX(layout, i, 0) + dx;
+                            MATRIX(*control_points, k, 1) = VECTOR(layer_to_y)[(igraph_integer_t) MATRIX(layout, i, 1)];
+                        }
+                    } else {
+                        for (k = 0; i < j; k++, i++) {
+                            MATRIX(*control_points, k, 0) = MATRIX(layout, i, 0) + dx;
+                            MATRIX(*control_points, k, 1) = VECTOR(layer_to_y)[(igraph_integer_t) MATRIX(layout, i, 1)];
+                        }
+                    }
+                }
+            }
+
+            /* Update the left margin for the next component */
+            max_x = 0;
+            for (i = 0; i < next_new_vertex_id; i++) {
+                if (MATRIX(layout, i, 0) > max_x) {
+                    max_x = MATRIX(layout, i, 0);
+                }
+            }
+            dx += max_x + hgap;
 
             igraph_destroy(&subgraph);
             igraph_i_layering_destroy(&layering);
@@ -538,23 +534,18 @@ igraph_error_t igraph_layout_sugiyama(const igraph_t *graph, igraph_matrix_t *re
         }
 
         igraph_vector_int_destroy(&new_layers);
+        igraph_vector_int_destroy(&new_vertex_id_to_edge_id);
         igraph_vector_int_destroy(&old2new_vertex_ids);
         igraph_vector_int_destroy(&new2old_vertex_ids);
         igraph_vector_int_destroy(&edgelist);
         igraph_vector_int_destroy(&neis);
-        IGRAPH_FINALLY_CLEAN(5);
+        IGRAPH_FINALLY_CLEAN(6);
     }
 
     igraph_vector_int_destroy(&layers_own);
     igraph_vector_destroy(&layer_to_y);
     igraph_vector_int_destroy(&membership);
     IGRAPH_FINALLY_CLEAN(3);
-
-    if (extd_graph != 0) {
-        IGRAPH_CHECK(igraph_create(extd_graph, &extd_edgelist, next_extd_vertex_id, igraph_is_directed(graph)));
-        igraph_vector_int_destroy(&extd_edgelist);
-        IGRAPH_FINALLY_CLEAN(1);
-    }
 
     return IGRAPH_SUCCESS;
 }